Update name and css for customer info component
