+++ practice-one/src/components/CustomerAvatar/CustomerInfo.css
@@ -6,7 +6,7 @@
 .customer__avatar {
   width: 30px;
   height: 30px;
-  border-radius: 100%;
+  border-radius: 50%;
   overflow: hidden;
   margin-right: 11px;
 }
